+++ src/app/manage-subject/manage-subject.component.ts
@@ -145,30 +145,45 @@ export class ManageSubjectComponent implements OnInit, OnDestroy {
     let subject : SubjectVM;
     let delsub : SubjectVM;
     delsub = subdata;
-    
-    this.confirmationService.confirm({
-      message: 'Are you sure that you want to delete',
-      accept: () => {
-        this.isloading = true;
-        subject = {
-          ...delsub,
-          isActive : false
-        }
-    
-        this.subs.sink = this.subjectservice.deleteSubject(subject).subscribe(data =>{
-          if(data){
-            this.subjectsAllData.forEach((element , index) => {
-              if(element.id === subdata.id){
-                this.subjectsAllData.splice(index , 1);
-                this.subjectTableData = this.subjectsAllData
-                this.isloading = false;
-                this.messageService.add({ severity: 'success', summary: 'Success', detail: 'Subject Removed' });
+
+    if(delsub.id && delsub.id>0 ){
+      this.subs.sink = this.subjectservice.checkGradeDeleteAvailability(delsub.id).subscribe(data =>{
+        if(data && data.content){
+          let isDeleteAvailable : Boolean = false;
+          isDeleteAvailable = data.content.isDeleteAvailable ? data.content.isDeleteAvailable : false;
+
+          if(isDeleteAvailable){
+            this.confirmationService.confirm({
+              message: 'Are you sure that you want to delete',
+              accept: () => {
+                this.isloading = true;
+                subject = {
+                  ...delsub,
+                  isActive : false
+                }
+            
+                this.subs.sink = this.subjectservice.deleteSubject(subject).subscribe(data =>{
+                  if(data){
+                    this.subjectsAllData.forEach((element , index) => {
+                      if(element.id === subdata.id){
+                        this.subjectsAllData.splice(index , 1);
+                        this.subjectTableData = this.subjectsAllData
+                        this.isloading = false;
+                        this.messageService.add({ severity: 'success', summary: 'Success', detail: 'Subject Removed' });
+                      }
+                    });
+                  }
+                })
               }
             });
+          }else{
+            this.messageService.add({ severity: 'warn', summary: 'Warning', detail: 'Delete Cannot Proceed.'});
           }
-        })
-      }
-  });
+        }
+      })
+    }
+    
+    
 
   }
